requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Database
supabase==2.4.3

# Data
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
cloudscraper==1.2.71

# ML
scikit-learn==1.4.2
catboost==1.2.5
joblib==1.4.0

# Dev
pytest==8.2.0
pytest-asyncio==0.23.6
black==24.4.2
catboost>=1.2
